refactor: organize views file
