+++ src/router/index.js
@@ -1,5 +1,4 @@
 import { createRouter, createWebHistory } from 'vue-router'
-import HomeView from '../views/HomeView.vue'
 
 const router = createRouter({
   history: createWebHistory(import.meta.env.BASE_URL),
@@ -7,48 +6,42 @@ const router = createRouter({
     {
       path: '/',
       name: 'home',
-      component: HomeView
+      component: () => import('../views/index.vue')
     },
     {
       path: '/about-us',
       name: 'about us',
-      // route level code-splitting
-      // this generates a separate chunk (About.[hash].js) for this route
-      // which is lazy-loaded when the route is visited.
       component: () => import('../views/about/index.vue')
     },
     {
       path: '/news-event',
       name: 'News & Events',
-      // route level code-splitting
-      // this generates a separate chunk (About.[hash].js) for this route
-      // which is lazy-loaded when the route is visited.
-      component: () => import('../views/NewsEvents.vue')
+      component: () => import('../views/news-event/index.vue')
     },
     {
       path: '/contact',
       name: 'contact',
-      component: () => import('../views/ContactView.vue')
+      component: () => import('../views/contact/index.vue')
     },
     {
       path: '/my-account',
       name: 'my account',
-      component: () => import('../views/MyAccountView.vue')
+      component: () => import('../views/my-account/index.vue')
     },
     {
       path: '/my-account/orders',
       name: 'Orders',
-      component: () => import('../views/OrdersView.vue')
+      component: () => import('../views/my-account/orders.vue')
     },
     {
       path: '/my-account/downloads',
       name: 'Downloads',
-      component: () => import('../views/DownloadsView.vue')
+      component: () => import('../views/my-account/downloads.vue')
     },
     {
       path: '/my-account/edit-address',
       name: 'Addresses',
-      component: () => import('../views/AddressesView.vue')
+      component: () => import('../views/addresses/index.vue')
     },
     {
       path: '/my-account/edit-address/billing',
@@ -58,32 +51,32 @@ const router = createRouter({
     {
       path: '/my-account/edit-address/shipping',
       name: 'Shipping Address',
-      component: () => import('../views/ShippingAddressView.vue')
+      component: () => import('../views/addresses/shipping.vue')
     },
     {
       path: '/my-account/edit-account',
       name: 'Account Details',
-      component: () => import('../views/AccountDetailsView.vue')
+      component: () => import('../views/my-account/details.vue')
     },
     {
       path: '/shop',
       name: 'Shop',
-      component: () => import('../views/ProductsView.vue')
+      component: () => import('../views/product/index.vue')
     },
     {
       path: '/product/:slug',
       name: 'Product Detail',
-      component: () => import('../views/ProductDetailView.vue')
+      component: () => import('../views/product/_slug/index.vue')
     },
     {
       path: '/detail-news-events/:id',
       name: 'DetailNewsEvents',
-      component: () => import('../views/DetailNewsEvents.vue')
+      component: () => import('../views/news-event/_slug/index.vue')
     },
     {
       path: '/cart',
       name: 'Cart',
-      component: () => import('../views/CartView.vue')
+      component: () => import('../views/cart/index.vue')
     }
   ]
 })
