+++ src/views/news-event/_slug/index.vue
@@ -1,7 +1,7 @@
 <script setup>
-import Navbar from '../components/Navbar.vue';
-import Footer from '../components/Footer.vue';
-import PostRecomendations from '../components/PostRecomendations.vue';
+import Navbar from '../../../components/Navbar.vue';
+import Footer from '../../../components/Footer.vue';
+import PostRecomendations from '../../../components/PostRecomendations.vue';
 
 const data = [
     {
